Re-structured the entirety of the GUI engine to be based on an event system, and removed Minecraft code as a dependency in GUI code.

+++ src/main/java/com/g4mesoft/gui/scroll/GSIScrollableViewport.java
@@ -16,11 +16,11 @@ public interface GSIScrollableViewport extends GSIViewport {
 		return getHeight();
 	}
 	
-	default public double getIncrementalScrollX(int sign) {
-		return Double.NaN;
+	default public float getIncrementalScrollX(int sign) {
+		return Float.NaN;
 	}
 
-	default public double getIncrementalScrollY(int sign) {
-		return Double.NaN;
+	default public float getIncrementalScrollY(int sign) {
+		return Float.NaN;
 	}
 }
